17_advance_agent/pyproject.toml
[project]
name = "17-advance-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.4.400",
    "openai-agents>=0.0.7",
    "python-dotenv>=1.1.0",
]
